requirements.txt
flask
line-bot-sdk
python-dotenv
requests
openai